+++ src/example_app/example_app/documents/__init__.py
@@ -16,13 +16,18 @@ def includeme(config: Configurator):
 
     def add_to_root(event):
         docs = event.root['docs'] = Docs()
-        for i in range(1, 1001):
+        for i in range(1, 1000005):
+            if i % 2 == 0:
+                tags = ['tag1', 'tag2']
+            else:
+                tags = ['tag3']
             docs.http_post(
                 event.request,
                 params={
                     'user_id': 1,
                     'name': f'Important document {i}',
                     'data': f'It is a data of document {i}.',
+                    'tags': tags,
                 },
             )